Raising the set_view_size to utils
